Suggestion for update to AuthRouteMethods
I appreciate all the work you and the team are doing. When upgrading from Laravel >= 5.7 and now 6.x to 7.x some of the methods and classes have changed. I kept getting a login and logout conflict with our web routes. I had to dig down and find this reference. The $options attribute is a great way to turn on/off routes but it seems a little incomplete. There are so many more suggestion dealing with following existing controller method naming conventions, should the options be in an auth configuration, etc. but wanted to focus on a low hanging fruit (this use case).

Use case: Web route name "login" and "logout" is defined but still want to use auth routes "registration" and "reset". Was able to turn on/off everything except "login" route.

I could see this being implemented in a minor version without backwards compatibility issues. Please consider.

Thanks,
Brant