Fix the issue of none existing directory
If we have the `js` directory in the `resources` the `bootstrap` file will be created successfully BUT what if the directory does not exist? The `copy(C:\...\resources\js/bootstrap.js): Failed to open stream: No such file or directory` exception will be triggered as happened WITH me So, this condition solve this issue, by checking if the directory was not existed then create it for me.