Use 'dashboard' route instead of 'home' route
We observed a modification in the Laravel project where the 'home' route was replaced with a 'dashboard' route in the welcome.blade.php file. As part of this update, we removed the 'home' route and added the new 'dashboard' URL to the package. All related files have been updated accordingly to reflect this change.